Kernel Scheduler: Extensive Audit Findings & Bug Squash (2025)
This commit addresses critical bugs and logic errors identified during
the 2025 extensive scheduler audit:

1. Fixed a memory corruption bug in RunQueue::Remove where raw 64-bit
   atomic AND was used on a native_cpu_mask_t field. On 32-bit systems
   this caused out-of-bounds writes. Replaced with architecture-aware
   cpu_mask_and_atomic helper.

2. Resolved a logic error in ThreadData::_ComputeEffectivePriority where
   overdue threads (negative deadline delta) incorrectly received
   minimum urgency due to unsigned wrap-around in fixed-point math.
   They now explicitly receive maximum dynamic priority.

3. Standardized atomic reads in low_latency.cpp and power_saving.cpp
   using LoadAcquire64 for layered flat bitmask discovery to ensure
   memory consistency across all supported architectures.

4. Updated scheduler_audit_2025.md and scheduler_audit_report_2025_final.md
   to document these findings as Task 9 of the roadmap.

5. Corrected indentation in CPUEntry::ComputeLoad.

+++ src/system/kernel/scheduler/scheduler_cpu.cpp
@@ -63,7 +63,7 @@ UpdateIdleCoreLFB(CoreEntry* core, bool idle)
 			uint64 summary = (uint64)LoadAcquire64(gIdleNodeSummary);
 			if (!(summary & (1ULL << nodeID))) break;
 
-			if (AtomicTestAndSet64(gIdleNodeSummary, (int64)(summary & ~(1ULL << nodeID)), (int64)summary) == (int64)summary)
+			if ((uint64)AtomicTestAndSet64(gIdleNodeSummary, (int64)(summary & ~(1ULL << nodeID)), (int64)summary) == summary)
 				break;
 		}
 	}
@@ -586,7 +586,7 @@ void CPUEntry::ComputeLoad(bigtime_t now) {
 							   tempLoad, now);
 		if (oldLoad < 0)
 			break;
-		if ((bigtime_t)AtomicTestAndSet64(fMeasureActiveTime, tempMeasureActiveTime, measureActiveTime) == measureActiveTime) {
+		if ((bigtime_t)AtomicTestAndSet64(fMeasureActiveTime, (int64)tempMeasureActiveTime, (int64)measureActiveTime) == measureActiveTime) {
 			StoreRelease64(fMeasureTime, tempMeasureTime);
 			currentLoad = tempLoad;
 			break;
